# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WadeCV API",
    "description": "AI-powered CV tailoring system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/jobs/gap-insights/": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Gap Insights",
        "description": "Return the logged-in user's gap insights, refreshing them every N applications.",
        "operationId": "get_gap_insights_api_jobs_gap_insights__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapInsightsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/jobs/": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "title": "Response List Jobs Api Jobs  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Jobs"
        ],
        "summary": "Update Job",
        "operationId": "update_job_api_jobs__job_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/scrape": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Scrape Job",
        "operationId": "scrape_job_api_jobs_scrape_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrapeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/mail/extract-address": {
      "post": {
        "tags": ["Physical Mail"],
        "summary": "Extract Address",
        "operationId": "extract_address_api_mail_extract_address_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractAddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractAddressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/mail/send": {
      "post": {
        "tags": ["Physical Mail"],
        "summary": "Send Mail",
        "operationId": "send_mail_api_mail_send_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_mail_api_mail_send_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhysicalMailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    }
  },
  "components": {
    "schemas": {
...
      "GapInsightTheme": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "label",
          "description",
          "count"
        ],
        "title": "GapInsightTheme"
      },
      "GapInsights": {
        "properties": {
          "summary_text": {
            "type": "string",
            "title": "Summary Text"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/GapInsightTheme"
            },
            "type": "array",
            "title": "Themes",
            "default": []
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "summary_text"
        ],
        "title": "GapInsights"
      },
      "GapInsightsResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "total_applications": {
            "type": "integer",
            "title": "Total Applications"
          },
          "next_refresh_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Refresh At"
          },
          "gap_insights": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GapInsights"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "available",
          "total_applications"
        ],
        "title": "GapInsightsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "cv_id": {
            "type": "string",
            "title": "Cv Id"
          },
          "job_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Url"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "application_status": {
            "type": "string",
            "title": "Application Status"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "fit_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cv_id",
          "job_url",
          "job_description",
          "company_name",
          "job_title",
          "application_status",
          "applied_at",
          "created_at"
        ],
        "title": "JobResponse"
      },
      "JobUpdateRequest": {
        "properties": {
          "application_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Status"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          }
        },
        "type": "object",
        "title": "JobUpdateRequest"
      },
...
      "ScrapeRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ScrapeRequest"
      },
      "ScrapeResponse": {
        "properties": {
          "job_description": {
            "type": "string",
            "title": "Job Description"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "job_description"
        ],
        "title": "ScrapeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ExtractAddressRequest": {
        "properties": {
          "job_id": { "type": "string", "title": "Job Id" }
        },
        "type": "object",
        "required": ["job_id"],
        "title": "ExtractAddressRequest"
      },
      "ExtractAddressResponse": {
        "properties": {
          "found": { "type": "boolean", "title": "Found" },
          "name": { "type": "string", "title": "Name", "default": "" },
          "address_line1": { "type": "string", "title": "Address Line1", "default": "" },
          "address_line2": { "type": "string", "title": "Address Line2", "default": "" },
          "city": { "type": "string", "title": "City", "default": "" },
          "state": { "type": "string", "title": "State", "default": "" },
          "zip": { "type": "string", "title": "Zip", "default": "" },
          "country": { "type": "string", "title": "Country", "default": "US" }
        },
        "type": "object",
        "required": ["found"],
        "title": "ExtractAddressResponse"
      },
      "Body_send_mail_api_mail_send_post": {
        "properties": {
          "job_id": { "type": "string", "title": "Job Id" },
          "content_type": { "type": "string", "enum": ["cv_only", "cover_letter_only", "both"], "title": "Content Type" },
          "to_address": { "type": "string", "title": "To Address", "description": "JSON-encoded address object" },
          "from_address": { "type": "string", "title": "From Address", "description": "JSON-encoded address object" },
          "save_return_address": { "type": "boolean", "title": "Save Return Address", "default": true },
          "custom_cv_pdf": { "type": "string", "contentMediaType": "application/octet-stream", "title": "Custom Cv Pdf" },
          "custom_cover_letter_pdf": { "type": "string", "contentMediaType": "application/octet-stream", "title": "Custom Cover Letter Pdf" }
        },
        "type": "object",
        "required": ["job_id", "content_type", "to_address", "from_address"],
        "title": "Body_send_mail_api_mail_send_post"
      },
      "PhysicalMailResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "job_id": { "type": "string", "title": "Job Id" },
          "lob_letter_id": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Lob Letter Id" },
          "content_type": { "type": "string", "title": "Content Type" },
          "status": { "type": "string", "title": "Status" },
          "credits_charged": { "type": "integer", "title": "Credits Charged" },
          "created_at": { "type": "string", "title": "Created At" }
        },
        "type": "object",
        "required": ["id", "job_id", "content_type", "status", "credits_charged", "created_at"],
        "title": "PhysicalMailResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}